Рефакторинг: - Все объекты распределены по пакетам. - Взаимодействие между слоем ui и слоем data происходит через слой domain посредством интерфейсов. - Вся логика плеера вынесена из AudioplayerActivity. Взаимодействие с плеером осуществляется через PlayerInteractor.
