+++ app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".App"
+        android:name=".app.App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -16,10 +16,10 @@
         android:theme="@style/Theme.PlaylistMaker"
         tools:targetApi="31" >
         <activity
-            android:name=".activities.AudioplayerActivity"
+            android:name=".presentaion.ui.audioplayer.AudioplayerActivity"
             android:exported="false" />
         <activity
-            android:name=".activities.SearchActivity"
+            android:name=".presentaion.ui.search.SearchActivity"
             android:exported="false"
             tools:ignore="LockedOrientationActivity" >
             <meta-data
@@ -27,7 +27,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".activities.MediaActivity"
+            android:name=".presentaion.ui.media.MediaActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity" >
@@ -36,7 +36,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".activities.SettingsActivity"
+            android:name=".presentaion.ui.settings.SettingsActivity"
             android:exported="false"
             android:theme="@style/Theme.PlaylistMaker.Settings"
             tools:ignore="LockedOrientationActivity" >
@@ -45,7 +45,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".activities.MainActivity"
+            android:name=".presentaion.ui.main_activity.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.PlaylistMaker.Main"
             tools:ignore="LockedOrientationActivity" >
