--- app/src/main/java/com/example/playlistmaker/activities/MainActivity.kt
@@ -1,29 +1,0 @@
-package com.example.playlistmaker.activities
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Button
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.R
-
-class MainActivity : AppCompatActivity() {
-    private val settingsButton by lazy { findViewById<Button>(R.id.settings) }
-    private val mediaButton by lazy { findViewById<Button>(R.id.media) }
-    private val searchButton by lazy { findViewById<Button>(R.id.search) }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        this.theme.changingConfigurations
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        searchButton.setOnClickListener {
-            startActivity(Intent(this, SearchActivity::class.java))
-        }
-        mediaButton.setOnClickListener {
-            startActivity(Intent(this, MediaActivity::class.java))
-        }
-        settingsButton.setOnClickListener {
-            startActivity(Intent(this, SettingsActivity::class.java))
-        }
-    }
-}
