--- app/src/main/java/com/example/playlistmaker/activities/MediaActivity.kt
@@ -1,13 +1,0 @@
-package com.example.playlistmaker.activities
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.R
-
-class MediaActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_media)
-    }
-
-}
