+++ app/src/main/java/com/example/playlistmaker/data/dto/TrackListDto.kt
@@ -1,0 +1,5 @@
+package com.example.playlistmaker.data.dto
+
+import com.example.playlistmaker.domain.entities.Track
+
+data class TrackListDto(val list: List<Track>)
