+++ app/src/main/java/com/example/playlistmaker/data/dto/TrackSearchRequest.kt
@@ -1,0 +1,3 @@
+package com.example.playlistmaker.data.dto
+
+data class TrackSearchRequest(val term: String)
