+++ app/src/main/java/com/example/playlistmaker/data/network/NetworkClient.kt
@@ -1,0 +1,5 @@
+package com.example.playlistmaker.data.network
+
+interface NetworkClient {
+    fun doRequest(dto: Any): Response
+}
