+++ app/src/main/java/com/example/playlistmaker/data/network/Response.kt
@@ -1,0 +1,3 @@
+package com.example.playlistmaker.data.network
+
+open class Response(var resultCode: Int = 0)
