+++ app/src/main/java/com/example/playlistmaker/domain/entities/Track.kt
@@ -1,0 +1,14 @@
+package com.example.playlistmaker.domain.entities
+
+data class Track(
+    val trackId: Long,
+    val trackName: String,
+    val artistName: String,
+    val trackTime: String,
+    val artworkUrl100: String,
+    val collectionName: String,
+    val country: String,
+    val releaseDate: String,
+    val primaryGenreName: String,
+    val previewUrl: String
+)
