+++ app/src/main/java/com/example/playlistmaker/domain/repositories/SearchRepository.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.domain.repositories
+
+import com.example.playlistmaker.domain.entities.Track
+
+interface SearchRepository {
+    fun search(term: String): List<Track>?
+}
