+++ app/src/main/java/com/example/playlistmaker/domain/repositories/StorageManagerRepo.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.domain.repositories
+
+interface StorageManagerRepo<T> {
+    fun save(key: String, item: T)
+
+    fun get(key: String): T
+}
