--- app/src/main/java/com/example/playlistmaker/itunes/ITunesResponse.kt
@@ -1,12 +1,0 @@
-package com.example.playlistmaker.itunes
-
-
-import com.example.playlistmaker.trackRecyclerView.Track
-import com.google.gson.annotations.SerializedName
-
-data class ITunesResponse(
-    @SerializedName("resultCount")
-    val resultCount: Int = 0,
-    @SerializedName("results")
-    val results: List<Track> = emptyList()
-)
