+++ app/src/main/java/com/example/playlistmaker/presentaion/ui/media/MediaActivity.kt
@@ -1,0 +1,13 @@
+package com.example.playlistmaker.presentaion.ui.media
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.playlistmaker.R
+
+class MediaActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_media)
+    }
+
+}
