+++ app/src/main/java/com/example/playlistmaker/presentaion/utils/Player.kt
@@ -1,0 +1,56 @@
+package com.example.playlistmaker.presentaion.utils
+
+import android.media.MediaPlayer
+
+class Player(
+    url: String,
+    onPrepared: Runnable,
+    onCopletion: Runnable
+) {
+    private val player = MediaPlayer()
+    private var playerState = STATE_DEFAULT
+
+
+    init {
+        player.setDataSource(url)
+        player.prepareAsync()
+        player.setOnPreparedListener {
+            playerState = STATE_PREPARED
+            onPrepared.run()
+        }
+        player.setOnCompletionListener {
+            playerState = STATE_PREPARED
+            onCopletion.run()
+        }
+    }
+
+    fun playbackControl() {
+        when (playerState) {
+            STATE_PREPARED, STATE_PAUSED -> playMedia()
+            STATE_PLAYING -> pauseMedia()
+        }
+    }
+
+    fun getCurrentPosition(): Int = player.currentPosition
+    private fun playMedia() {
+        player.start()
+        playerState = STATE_PLAYING
+    }
+
+    private fun pauseMedia() {
+        player.pause()
+        playerState = STATE_PAUSED
+    }
+
+    fun releaseResources() {
+        player.release()
+        playerState = STATE_DEFAULT
+    }
+
+    companion object {
+        private const val STATE_DEFAULT = 0
+        private const val STATE_PREPARED = 1
+        private const val STATE_PLAYING = 2
+        private const val STATE_PAUSED = 3
+    }
+}
