--- app/src/main/java/com/example/playlistmaker/trackRecyclerView/TrackAdapter.kt
@@ -1,24 +1,0 @@
-package com.example.playlistmaker.trackRecyclerView
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import com.example.playlistmaker.R
-
-class TrackAdapter(private val onTrackClicked: (Track) -> Unit) : RecyclerView.Adapter<TrackViewHolder>() {
-    var trackList: MutableList<Track> = mutableListOf()
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
-        return TrackViewHolder(LayoutInflater.from(parent.context).inflate(R.layout.track_card, parent, false))
-    }
-
-    override fun getItemCount(): Int {
-        return trackList.size
-    }
-
-    override fun onBindViewHolder(holder: TrackViewHolder, position: Int) {
-        val track = trackList[position]
-        holder.bind(track)
-        holder.itemView.setOnClickListener { onTrackClicked(track) }
-    }
-}
