+++ app/src/main/res/layout/activity_audioplayer.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     android:orientation="vertical"
-    tools:context=".activities.AudioplayerActivity">
+    tools:context=".presentaion.ui.audioplayer.AudioplayerActivity">
     <!-- Back -->
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
@@ -140,6 +140,7 @@
                 app:layout_constraintBottom_toBottomOf="@id/add_to_media"
                 app:layout_constraintStart_toEndOf="@id/left_border"
                 app:layout_constraintEnd_toStartOf="@id/right_border"
+                android:clickable="false"
                 android:src="@drawable/play_icon"
                 app:tint="?attr/colorOnSecondary" />
             <ImageButton
