+++ app/src/main/res/layout/activity_media.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.MediaActivity">
+    tools:context=".presentaion.ui.media.MediaActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
