+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@ xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 android:orientation="vertical"
-tools:context=".activities.SearchActivity" >
+tools:context=".presentaion.ui.search.SearchActivity" >
 
     <!-- Head -->
     <com.google.android.material.appbar.MaterialToolbar
